float cast de float cast

+++ Assets/Code/Statics/ActionStatics.cs
@@ -248,7 +248,7 @@ public static class ActionStatics
     {
         UserBattler.statusmultiplier = StatusCalculation(UserBattler, TargetBattler);
 
-        UserBattler.luckymultiplier = Mathf.Pow(1.5, ((int) (UserBattler.luckp/60)));
+        UserBattler.luckymultiplier = (float) Mathf.Pow(1.5f, (float) ((int) (UserBattler.luckp/60)));
         
         //Damage = (((ATK + ATK BOOST) x ATK MULTIPLIER) x LUCKY - DEF + DMG Buffs) x CRIT x STATUS MULTIPLIER x GUARD MULTIPLIER + FLAT DAMAGE/MINIMUM DAMAGE
         int DMG = (int)((((UserBattler.atk + UserBattler.atkboost) * UserBattler.atkmultiplier) * UserBattler.luckymultiplier - ((TargetBattler.def + TargetBattler.defboost) * TargetBattler.defmultiplier) + UserBattler.dmgboost) * UserBattler.critmultiplier * UserBattler.statusmultiplier * TargetBattler.guardmultiplier);
